pyproject.toml
[build-system]
requires = ["setuptools>=70", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "deft-indexer"
version = "0.1.0"
description = "Prototype blockchain event indexer with FastAPI, Celery, PostgreSQL, and Kafka"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.27",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.32",
    "sqlalchemy>=2.0",
    "psycopg[binary]>=3.2",
    "pydantic>=2.5",
    "pydantic-settings>=2.3",
    "celery>=5.4",
    "redis>=5.0",
    "web3>=7.4",
    "kafka-python-ng>=2.2",
    "python-dotenv>=1.0",
    "structlog>=24.1",
    "prometheus-client>=0.20"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.24",
    "httpx>=0.27",
    "asgi-lifespan>=2.1"
]

[tool.pytest.ini_options]
asyncio_mode = "strict"
testpaths = ["tests"]
